# docs/openapi.yaml
openapi: 3.0.0
info:
  title: SaaS Boilerplate API
  description: Multi-tenant SaaS platform API for managing tenants, organizations, users, and roles
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@yourdomain.com
    url: https://yourdomain.com/support
  license:
    name: MIT
    url: https://github.com/your-org/saas-boilerplate/blob/main/LICENSE
servers:
  - url: https://api.yourdomain.com/v1
    description: Production server
  - url: https://staging-api.yourdomain.com/v1
    description: Staging server
  - url: http://localhost:3000/v1
    description: Development server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: '550e8400-e29b-41d4-a716-446655440002'
        tenantId:
          type: string
          format: uuid
          example: '550e8400-e29b-41d4-a716-446655440000'
        organizationId:
          type: string
          format: uuid
          nullable: true
          example: '550e8400-e29b-41d4-a716-446655440001'
        email:
          type: string
          format: email
          example: 'user@example.com'
        firstName:
          type: string
          minLength: 1
          maxLength: 100
          nullable: true
          example: 'John'
        lastName:
          type: string
          minLength: 1
          maxLength: 100
          nullable: true
          example: 'Doe'
        status:
          type: string
          enum: [active, inactive, suspended]
          example: 'active'
        emailVerified:
          type: boolean
          example: true
        preferences:
          type: object
          additionalProperties: true
          example:
            theme: 'dark'
            notifications: true
        metadata:
          type: object
          additionalProperties: true
          example:
            source: 'api'
            role: 'developer'
        createdAt:
          type: string
          format: date-time
          example: '2024-01-01T00:00:00.000Z'
        updatedAt:
          type: string
          format: date-time
          example: '2024-01-01T00:00:00.000Z'
      required:
        - id
        - tenantId
        - email
        - status
        - createdAt
        - updatedAt
...
    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
        total:
          type: integer
          minimum: 0
          example: 150
        page:
          type: integer
          minimum: 1
          example: 2
        limit:
          type: integer
          minimum: 1
          maximum: 100
          example: 20
        hasNext:
          type: boolean
          example: true
        hasPrev:
          type: boolean
          example: true
      required:
        - data
        - total
        - page
        - limit
        - hasNext
        - hasPrev
...
    CreateUserRequest:
      type: object
      properties:
        tenantId:
          type: string
          format: uuid
          example: '550e8400-e29b-41d4-a716-446655440000'
        organizationId:
          type: string
          format: uuid
          example: '550e8400-e29b-41d4-a716-446655440001'
        email:
          type: string
          format: email
          example: 'user@example.com'
        firstName:
          type: string
          minLength: 1
          maxLength: 100
          example: 'John'
        lastName:
          type: string
          minLength: 1
          maxLength: 100
          example: 'Doe'
        roleIds:
          type: array
          items:
            type: string
            format: uuid
          example: ['550e8400-e29b-41d4-a716-446655440003']
        preferences:
          type: object
          additionalProperties: true
          example:
            theme: 'dark'
            notifications: true
        metadata:
          type: object
          additionalProperties: true
          example:
            source: 'api'
            role: 'developer'
      required:
        - tenantId
        - organizationId
        - email
        - firstName
        - lastName
...
  parameters:
    TenantIdParam:
      name: tenantId
      in: query
      description: Filter results by tenant ID
      required: false
      schema:
        type: string
        format: uuid
        example: '550e8400-e29b-41d4-a716-446655440000'
    PageParam:
      name: page
      in: query
      description: Page number for pagination
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
        example: 1
    LimitParam:
      name: limit
      in: query
      description: Number of items per page
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
        example: 20
    SearchParam:
      name: search
      in: query
      description: Search term for filtering results
      required: false
      schema:
        type: string
        example: 'demo'
    StatusParam:
      name: status
      in: query
      description: Filter by status
      required: false
      schema:
        type: string
        enum: [active, inactive, suspended]
        example: 'active'
    IdPathParam:
      name: id
      in: path
      description: Resource ID
      required: true
      schema:
        type: string
        format: uuid
        example: '550e8400-e29b-41d4-a716-446655440000'
paths:
...
  /users:
    get:
      summary: List users
      description: Retrieve a paginated list of users for a specific tenant
      tags:
        - Users
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantIdParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SearchParam'
        - $ref: '#/components/parameters/StatusParam'
      responses:
        '200':
          description: List of users retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create user
      description: Create a new user within a tenant and organization
      tags:
        - Users
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /users/{id}:
    get:
      summary: Get user by ID
      description: Retrieve a specific user by their ID
      tags:
        - Users
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPathParam'
        - $ref: '#/components/parameters/TenantIdParam'
      responses:
        '200':
          description: User retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...